Patch up scope leak with a resolver static pass
The scope leak introduced in the previous commit with closures has to do with the mutability of the environment that is passed by reference to the closure.

By modifying that environment later outside of a function, you also end up mutating the closure, causing a violation of Lox's principle of static scope.

To fix this, semantic analysis is done within a resolver to capture at compile-time a set number of "hops" from the scope of any variable expression or assignment to the environment of the scope which contains that variable's declaration.

This allows not only for more efficient lookup of variable declarations at runtime, but patches the scope leak since every variable expression / assignment will always point to the same correct environment which contains its original binding.